fix(kiwi-cli-tools): changed directory of old "tools".
