+++ packages/kiwi_generator/pubspec.yaml
@@ -16,7 +16,7 @@ dependencies:
   code_builder: ^4.10.0
   dart_style: ^2.2.4
   kiwi: ^4.1.0
-  path: ^1.9.0
+  path: ^1.8.3
   source_gen: ^1.5.0
 
 dev_dependencies:
