--- tools/actions/pubspec.yaml
@@ -1,11 +1,0 @@
-name: actions_tools
-description: A tool for the github actions.
-version: 1.0.0+2
-publish_to: none
-
-environment:
-  sdk: '>=3.0.0 <4.0.0'
-
-dev_dependencies:
-  test: ^1.23.1
-  yaml: ^3.1.1
